Get emails from click house v0

+++ dags/oss_know/libs/github/sync_profiles.py
@@ -4,7 +4,7 @@ import requests
 from oss_know.libs.util.base import get_opensearch_client, get_redis_client, \
     infer_country_company_geo_insert_into_profile
 from oss_know.libs.util.github_api import GithubAPI
-from oss_know.libs.base_dict.opensearch_index import OPENSEARCH_INDEX_GITHUB_PROFILE, OPENSEARCH_INDEX_CHECK_SYNC_DATA
+from oss_know.libs.base_dict.opensearch_index import OPENSEARCH_INDEX_GITHUB_PROFILE
 from oss_know.libs.base_dict.redis import STORAGE_HASH, STORAGE_IDS_SET, WORKING_HASH
 from oss_know.libs.util.log import logger
 from opensearchpy import helpers as opensearch_helpers
@@ -93,7 +93,7 @@ def sync_github_profiles(github_tokens, opensearch_conn_info, redis_client_info,
                 infer_country_company_geo_insert_into_profile(latest_github_profile)
                 opensearch_client.index(index=OPENSEARCH_INDEX_GITHUB_PROFILE,
                                         body={"search_key": {
-                                            'updated_at': (datetime.datetime.now().timestamp() * 1000)},
+                                            'updated_at': int(datetime.datetime.now().timestamp() * 1000)},
                                             "raw_data": latest_github_profile},
                                         refresh=True)
                 logger.info(f"Success put updated {storage_id}'s github profiles into opensearch.")
